requirements.txt
playwright
beautifulsoup4
jinja2
python-dotenv
requests
flask

openai>=1.0.0
